pretty sure I fixed polygons
There were a couple issues with my project-then-clip strategy:

1. faces in 3-Space which were backwards from eye POV would not render.

This isn't an issue for solid objects, but if I maybe wanted to get the
exit point & normal, the code would not detect hits with the backface.

This occured because I was supplying the clockwise normal to a pair of
vertices for clipping, this is an issue for back faces.

This was fixed by checking the sign of perp_{plane}(normal) (where the
normal refers to the 3D normal, and the plane is the projected-onto
plane).

If the the sign was positive, then we do CCW clipping (as
normal) otherwise we do CW clipping.

2. when projecting onto the xz plane, the coordinates were flipped the
wrong way

I was going from (x, y, z) -> (x, z), I needed to go from (x, y, z)
-> (z, x) to mainain that -y faced forwards.